draft of modify
