--- bridgectrlv2/config.go
@@ -1,9 +1,0 @@
-package bridgectrlv2
-
-// Config is state config
-type Config struct {
-	// Store is the kind of storage in the bridge tree
-	Store string
-	// Height is the depth of the merkle tree
-	Height uint8
-}
