+++ db/storage_test.go
@@ -4,7 +4,6 @@ import (
 	"context"
 	"math/big"
 	"testing"
-	"time"
 
 	"github.com/0xPolygonHermez/zkevm-bridge-service/db/pgstorage"
 	"github.com/0xPolygonHermez/zkevm-bridge-service/etherman"
@@ -33,7 +32,7 @@ func TestExitRootStore(t *testing.T) {
 	storage, err := NewStorage(storageCfg, networksNumber)
 	require.NoError(t, err)
 	var networkID uint = 1
-	_, err = storage.GetLatestL1SyncedExitRoot(ctx)
+	_, err = storage.GetLatestL1SyncedExitRoot(ctx, nil)
 	require.Error(t, err)
 
 	var exitRoot etherman.GlobalExitRoot
@@ -45,13 +44,13 @@ func TestExitRootStore(t *testing.T) {
 	block.BlockNumber = 1
 	block.BlockHash = common.HexToHash("0x29e885edaf8e4b51e1d2e05f9da28161d2fb4f6b1d53827d9b80a23cf2d7d9fc")
 	block.NetworkID = networkID
-	id, err := storage.AddBlock(ctx, &block)
+	id, err := storage.AddBlock(ctx, &block, nil)
 	require.NoError(t, err)
 	exitRoot.BlockID = id
-	err = storage.AddExitRoot(ctx, &exitRoot)
+	err = storage.AddGlobalExitRoot(ctx, &exitRoot, nil)
 	require.NoError(t, err)
 
-	exit, err := storage.GetLatestExitRoot(ctx)
+	exit, err := storage.GetLatestExitRoot(ctx, nil)
 	require.NoError(t, err)
 	assert.Equal(t, exitRoot.GlobalExitRootNum, exit.GlobalExitRootNum)
 	assert.Equal(t, exitRoot.ExitRoots[0], exit.ExitRoots[0])
@@ -70,10 +69,10 @@ func TestExitRootStore(t *testing.T) {
 		NetworkID:          2,
 		TxHash:             common.HexToHash("0xc67996b03ac2ca401822b5be568e828f432121dd07ddedb306d9203e67675db8"),
 	}
-	err = storage.AddClaim(ctx, &claim)
+	err = storage.AddClaim(ctx, &claim, nil)
 	require.NoError(t, err)
 
-	claimStored, err := storage.GetClaim(ctx, claim.Index, claim.NetworkID)
+	claimStored, err := storage.GetClaim(ctx, claim.Index, claim.NetworkID, nil)
 	require.NoError(t, err)
 	assert.Equal(t, claim.Amount, claimStored.Amount)
 	assert.Equal(t, claim.BlockNumber, claimStored.BlockNumber)
@@ -95,10 +94,10 @@ func TestExitRootStore(t *testing.T) {
 		BlockID:            id,
 		TxHash:             common.HexToHash("0xc67996b03ac2ca401822b5be568e828f432121dd07ddedb306d9203e67675db8"),
 	}
-	err = storage.AddDeposit(ctx, &deposit)
+	err = storage.AddDeposit(ctx, &deposit, nil)
 	require.NoError(t, err)
 
-	depositStored, err := storage.GetDeposit(ctx, deposit.DepositCount, deposit.NetworkID)
+	depositStored, err := storage.GetDeposit(ctx, deposit.DepositCount, deposit.NetworkID, nil)
 	require.NoError(t, err)
 	assert.Equal(t, deposit.Amount, depositStored.Amount)
 	assert.Equal(t, deposit.BlockNumber, depositStored.BlockNumber)
@@ -117,53 +116,13 @@ func TestExitRootStore(t *testing.T) {
 		BlockID:              id,
 		NetworkID:            2,
 	}
-	err = storage.AddTokenWrapped(ctx, &tokenWrapped)
+	err = storage.AddTokenWrapped(ctx, &tokenWrapped, nil)
 	require.NoError(t, err)
 
-	tokenWrappedStored, err := storage.GetTokenWrapped(ctx, tokenWrapped.OriginalNetwork, tokenWrapped.OriginalTokenAddress)
+	tokenWrappedStored, err := storage.GetTokenWrapped(ctx, tokenWrapped.OriginalNetwork, tokenWrapped.OriginalTokenAddress, nil)
 	require.NoError(t, err)
 	assert.Equal(t, tokenWrapped.BlockNumber, tokenWrappedStored.BlockNumber)
 	assert.Equal(t, tokenWrapped.OriginalNetwork, tokenWrappedStored.OriginalNetwork)
 	assert.Equal(t, tokenWrapped.OriginalTokenAddress, tokenWrappedStored.OriginalTokenAddress)
 	assert.Equal(t, tokenWrapped.WrappedTokenAddress, tokenWrappedStored.WrappedTokenAddress)
-
-	// Batch
-	batch := etherman.Batch{
-		BlockNumber:    1,
-		Sequencer:      common.HexToAddress("0x29e885edaf8e4b51e1d2e05f9da28161d2fb4f6b1d53827d9b80a23cf2d7d9fe"),
-		ChainID:        big.NewInt(100),
-		GlobalExitRoot: common.HexToHash("0x30e885edaf8e4b51e1d2e05f9da28161d2fb4f6b1d53827d9b80a23cf2d7d9fe"),
-		TxHash:         common.Hash{},
-		BatchNumber:    1,
-		ReceivedAt:     time.Now(),
-		BlockID:        id,
-		NetworkID:      networkID,
-	}
-	err = storage.AddBatch(ctx, &batch)
-	require.NoError(t, err)
-
-	batchStored, err := storage.GetBatchByNumber(ctx, batch.BatchNumber, networkID)
-	require.NoError(t, err)
-	assert.Equal(t, batch.BlockNumber, batchStored.BlockNumber)
-	assert.Equal(t, batch.Sequencer, batchStored.Sequencer)
-	assert.Equal(t, batch.ChainID, batchStored.ChainID)
-	assert.Equal(t, batch.GlobalExitRoot, batchStored.GlobalExitRoot)
-	assert.Equal(t, batch.BatchNumber, batchStored.BatchNumber)
-	assert.Equal(t, common.Hash{}, batchStored.ConsolidatedTxHash)
-
-	batch.Aggregator = common.HexToAddress("0x29e885edaf8e4b51e1d2e05f9da28161d2fb4f6b1d53827d9b80a23cf2d7d9fc")
-	ti := time.Now()
-	batch.ConsolidatedAt = &ti
-	batch.ConsolidatedTxHash = common.HexToHash("0x31e885edaf8e4b51e1d2e05f9da28161d2fb4f6b1d53827d9b80a23cf2d7d9fe")
-	err = storage.ConsolidateBatch(ctx, &batch)
-	require.NoError(t, err)
-	batchStored2, err := storage.GetBatchByNumber(ctx, batch.BatchNumber, networkID)
-	require.NoError(t, err)
-	assert.Equal(t, batch.BlockNumber, batchStored2.BlockNumber)
-	assert.Equal(t, batch.Sequencer, batchStored2.Sequencer)
-	assert.Equal(t, batch.ChainID, batchStored2.ChainID)
-	assert.Equal(t, batch.GlobalExitRoot, batchStored2.GlobalExitRoot)
-	assert.Equal(t, batch.BatchNumber, batchStored2.BatchNumber)
-	assert.Equal(t, common.HexToHash("0x31e885edaf8e4b51e1d2e05f9da28161d2fb4f6b1d53827d9b80a23cf2d7d9fe"), batchStored2.ConsolidatedTxHash)
-	assert.Equal(t, common.HexToAddress("0x29e885edaf8e4b51e1d2e05f9da28161d2fb4f6b1d53827d9b80a23cf2d7d9fc"), batchStored2.Aggregator)
 }
