+++ etherman/config.go
@@ -2,6 +2,7 @@ package etherman
 
 // Config represents the configuration of the etherman
 type Config struct {
+	URL    string   `mapstructure:"URL"`
 	L1URL  string   `mapstructure:"L1URL"`
 	L2URLs []string `mapstructure:"L2URLs"`
