--- ethermanv2/config.go
@@ -1,6 +1,0 @@
-package ethermanv2
-
-// Config represents the configuration of the etherman
-type Config struct {
-	URL string `mapstructure:"URL"`
-}
