--- ethermanv2/etherman.go
@@ -1,628 +1,0 @@
-package ethermanv2
-
-import (
-	"context"
-	"encoding/json"
-	"errors"
-	"fmt"
-	"math/big"
-	"strings"
-	"time"
-
-	"github.com/0xPolygonHermez/zkevm-node/ethermanv2/smartcontracts/bridge"
-	"github.com/0xPolygonHermez/zkevm-node/ethermanv2/smartcontracts/globalexitrootmanager"
-	"github.com/0xPolygonHermez/zkevm-node/ethermanv2/smartcontracts/proofofefficiency"
-	"github.com/0xPolygonHermez/zkevm-node/log"
-	"github.com/ethereum/go-ethereum"
-	"github.com/ethereum/go-ethereum/accounts/abi"
-	"github.com/ethereum/go-ethereum/accounts/abi/bind"
-	"github.com/ethereum/go-ethereum/common"
-	"github.com/ethereum/go-ethereum/core/types"
-	"github.com/ethereum/go-ethereum/crypto"
-	"github.com/ethereum/go-ethereum/ethclient"
-	"golang.org/x/crypto/sha3"
-)
-
-var (
-	ownershipTransferredSignatureHash  = crypto.Keccak256Hash([]byte("OwnershipTransferred(address,address)"))
-	updateGlobalExitRootSignatureHash  = crypto.Keccak256Hash([]byte("UpdateGlobalExitRoot(uint256,bytes32,bytes32)"))
-	forcedBatchSignatureHash           = crypto.Keccak256Hash([]byte("ForceBatch(uint64,bytes32,address,bytes)"))
-	sequencedBatchesEventSignatureHash = crypto.Keccak256Hash([]byte("SequenceBatches(uint64)"))
-	forceSequencedBatchesSignatureHash = crypto.Keccak256Hash([]byte("SequenceForceBatches(uint64)"))
-	verifyBatchSignatureHash           = crypto.Keccak256Hash([]byte("VerifyBatch(uint64,address)"))
-	depositEventSignatureHash          = crypto.Keccak256Hash([]byte("BridgeEvent(uint32,address,uint32,address,uint256,bytes,uint32)"))
-	claimEventSignatureHash            = crypto.Keccak256Hash([]byte("ClaimEvent(uint32,uint32,address,address,uint256)"))
-	newWrappedTokenEventSignatureHash  = crypto.Keccak256Hash([]byte("NewWrappedToken(uint32,address,address)"))
-
-	// ErrNotFound is used when the object is not found
-	ErrNotFound = errors.New("Not found")
-)
-
-// EventOrder is the the type used to identify the events order
-type EventOrder string
-
-const (
-	// GlobalExitRootsOrder identifies a GlobalExitRoot event
-	GlobalExitRootsOrder EventOrder = "GlobalExitRoot"
-	// SequenceBatchesOrder identifies a VerifyBatch event
-	SequenceBatchesOrder EventOrder = "SequenceBatches"
-	// VerifyBatchOrder identifies a VerifyBatch event
-	VerifyBatchOrder EventOrder = "VerifyBatch"
-	// SequenceForceBatchesOrder identifies a SequenceForceBatches event
-	SequenceForceBatchesOrder EventOrder = "SequenceForceBatches"
-	// ForcedBatchesOrder identifies a ForcedBatches event
-	ForcedBatchesOrder EventOrder = "ForcedBatches"
-	// DepositsOrder identifies a Deposits event
-	DepositsOrder EventOrder = "Deposit"
-	// ClaimsOrder identifies a Claims event
-	ClaimsOrder EventOrder = "Claim"
-	// TokensOrder identifies a TokenWrapped event
-	TokensOrder EventOrder = "TokenWrapped"
-)
-
-type ethClienter interface {
-	ethereum.ChainReader
-	ethereum.LogFilterer
-	ethereum.TransactionReader
-}
-
-// Client is a simple implementation of EtherMan.
-type Client struct {
-	EtherClient           ethClienter
-	PoE                   *proofofefficiency.Proofofefficiency
-	Bridge                *bridge.Bridge
-	GlobalExitRootManager *globalexitrootmanager.Globalexitrootmanager
-	SCAddresses           []common.Address
-
-	auth *bind.TransactOpts
-}
-
-// NewClient creates a new etherman.
-func NewClient(cfg Config, auth *bind.TransactOpts, PoEAddr, bridgeAddr, globalExitRootManAddr common.Address) (*Client, error) {
-	// Connect to ethereum node
-	ethClient, err := ethclient.Dial(cfg.URL)
-	if err != nil {
-		log.Errorf("error connecting to %s: %+v", cfg.URL, err)
-		return nil, err
-	}
-	// Create smc clients
-	poe, err := proofofefficiency.NewProofofefficiency(PoEAddr, ethClient)
-	if err != nil {
-		return nil, err
-	}
-	bridge, err := bridge.NewBridge(bridgeAddr, ethClient)
-	if err != nil {
-		return nil, err
-	}
-	globalExitRoot, err := globalexitrootmanager.NewGlobalexitrootmanager(globalExitRootManAddr, ethClient)
-	if err != nil {
-		return nil, err
-	}
-	var scAddresses []common.Address
-	scAddresses = append(scAddresses, PoEAddr, globalExitRootManAddr, bridgeAddr)
-
-	return &Client{EtherClient: ethClient, PoE: poe, Bridge: bridge, GlobalExitRootManager: globalExitRoot, SCAddresses: scAddresses, auth: auth}, nil
-}
-
-// GetRollupInfoByBlockRange function retrieves the Rollup information that are included in all this ethereum blocks
-// from block x to block y.
-func (etherMan *Client) GetRollupInfoByBlockRange(ctx context.Context, fromBlock uint64, toBlock *uint64) ([]Block, map[common.Hash][]Order, error) {
-	// Filter query
-	query := ethereum.FilterQuery{
-		FromBlock: new(big.Int).SetUint64(fromBlock),
-		Addresses: etherMan.SCAddresses,
-	}
-	if toBlock != nil {
-		query.ToBlock = new(big.Int).SetUint64(*toBlock)
-	}
-	blocks, blocksOrder, err := etherMan.readEvents(ctx, query)
-	if err != nil {
-		return nil, nil, err
-	}
-	return blocks, blocksOrder, nil
-}
-
-// Order contains the event order to let the synchronizer store the information following this order.
-type Order struct {
-	Name EventOrder
-	Pos  int
-}
-
-func (etherMan *Client) readEvents(ctx context.Context, query ethereum.FilterQuery) ([]Block, map[common.Hash][]Order, error) {
-	logs, err := etherMan.EtherClient.FilterLogs(ctx, query)
-	if err != nil {
-		return nil, nil, err
-	}
-	var blocks []Block
-	blocksOrder := make(map[common.Hash][]Order)
-	for _, vLog := range logs {
-		err := etherMan.processEvent(ctx, vLog, &blocks, &blocksOrder)
-		if err != nil {
-			log.Warnf("error processing event. Retrying... Error: %s. vLog: %+v", err.Error(), vLog)
-			return nil, nil, err
-		}
-	}
-	return blocks, blocksOrder, nil
-}
-
-func (etherMan *Client) processEvent(ctx context.Context, vLog types.Log, blocks *[]Block, blocksOrder *map[common.Hash][]Order) error {
-	switch vLog.Topics[0] {
-	case sequencedBatchesEventSignatureHash:
-		return etherMan.sequencedBatchesEvent(ctx, vLog, blocks, blocksOrder)
-	case ownershipTransferredSignatureHash:
-		return etherMan.ownershipTransferredEvent(vLog)
-	case updateGlobalExitRootSignatureHash:
-		return etherMan.updateGlobalExitRootEvent(ctx, vLog, blocks, blocksOrder)
-	case forcedBatchSignatureHash:
-		return etherMan.forcedBatchEvent(ctx, vLog, blocks, blocksOrder)
-	case verifyBatchSignatureHash:
-		return etherMan.verifyBatchEvent(ctx, vLog, blocks, blocksOrder)
-	case forceSequencedBatchesSignatureHash:
-		return etherMan.forceSequencedBatchesEvent(ctx, vLog, blocks, blocksOrder)
-	case depositEventSignatureHash:
-		return etherMan.depositEvent(ctx, vLog, blocks, blocksOrder)
-	case claimEventSignatureHash:
-		return etherMan.claimEvent(ctx, vLog, blocks, blocksOrder)
-	case newWrappedTokenEventSignatureHash:
-		return etherMan.tokenWrappedEvent(ctx, vLog, blocks, blocksOrder)
-	}
-	log.Warn("Event not registered: ", vLog)
-	return nil
-}
-
-func (etherMan *Client) ownershipTransferredEvent(vLog types.Log) error {
-	ownership, err := etherMan.Bridge.ParseOwnershipTransferred(vLog)
-	if err != nil {
-		return err
-	}
-	emptyAddr := common.Address{}
-	if ownership.PreviousOwner == emptyAddr {
-		log.Debug("New rollup smc deployment detected. Deployment account: ", ownership.NewOwner)
-	} else {
-		log.Debug("Rollup smc OwnershipTransferred from account ", ownership.PreviousOwner, " to ", ownership.NewOwner)
-	}
-	return nil
-}
-
-func (etherMan *Client) updateGlobalExitRootEvent(ctx context.Context, vLog types.Log, blocks *[]Block, blocksOrder *map[common.Hash][]Order) error {
-	log.Debug("UpdateGlobalExitRoot event detected")
-	globalExitRoot, err := etherMan.GlobalExitRootManager.ParseUpdateGlobalExitRoot(vLog)
-	if err != nil {
-		return err
-	}
-	var gExitRoot GlobalExitRoot
-	gExitRoot.ExitRoots = make([]common.Hash, 0)
-	gExitRoot.ExitRoots = append(gExitRoot.ExitRoots, common.BytesToHash(globalExitRoot.MainnetExitRoot[:]))
-	gExitRoot.ExitRoots = append(gExitRoot.ExitRoots, common.BytesToHash(globalExitRoot.RollupExitRoot[:]))
-	gExitRoot.GlobalExitRootNum = globalExitRoot.GlobalExitRootNum
-	gExitRoot.GlobalExitRoot = hash(globalExitRoot.MainnetExitRoot, globalExitRoot.RollupExitRoot)
-	gExitRoot.BlockNumber = vLog.BlockNumber
-
-	if len(*blocks) == 0 || ((*blocks)[len(*blocks)-1].BlockHash != vLog.BlockHash || (*blocks)[len(*blocks)-1].BlockNumber != vLog.BlockNumber) {
-		fullBlock, err := etherMan.EtherClient.BlockByHash(ctx, vLog.BlockHash)
-		if err != nil {
-			return fmt.Errorf("error getting hashParent. BlockNumber: %d. Error: %w", vLog.BlockNumber, err)
-		}
-		block := prepareBlock(vLog, time.Unix(int64(fullBlock.Time()), 0), fullBlock)
-		block.GlobalExitRoots = append(block.GlobalExitRoots, gExitRoot)
-		*blocks = append(*blocks, block)
-	} else if (*blocks)[len(*blocks)-1].BlockHash == vLog.BlockHash && (*blocks)[len(*blocks)-1].BlockNumber == vLog.BlockNumber {
-		(*blocks)[len(*blocks)-1].GlobalExitRoots = append((*blocks)[len(*blocks)-1].GlobalExitRoots, gExitRoot)
-	} else {
-		log.Error("Error processing UpdateGlobalExitRoot event. BlockHash:", vLog.BlockHash, ". BlockNumber: ", vLog.BlockNumber)
-		return fmt.Errorf("error processing UpdateGlobalExitRoot event")
-	}
-	or := Order{
-		Name: GlobalExitRootsOrder,
-		Pos:  len((*blocks)[len(*blocks)-1].GlobalExitRoots) - 1,
-	}
-	(*blocksOrder)[(*blocks)[len(*blocks)-1].BlockHash] = append((*blocksOrder)[(*blocks)[len(*blocks)-1].BlockHash], or)
-	return nil
-}
-
-func (etherMan *Client) depositEvent(ctx context.Context, vLog types.Log, blocks *[]Block, blocksOrder *map[common.Hash][]Order) error {
-	log.Debug("Deposit event detected")
-	d, err := etherMan.Bridge.ParseBridgeEvent(vLog)
-	if err != nil {
-		return err
-	}
-	var deposit Deposit
-	deposit.Amount = d.Amount
-	deposit.BlockNumber = vLog.BlockNumber
-	deposit.OriginalNetwork = uint(d.OriginNetwork)
-	deposit.DestinationAddress = d.DestinationAddress
-	deposit.DestinationNetwork = uint(d.DestinationNetwork)
-	deposit.TokenAddress = d.OriginTokenAddress
-	deposit.DepositCount = uint(d.DepositCount)
-	deposit.TxHash = vLog.TxHash
-	deposit.Metadata = d.Metadata
-
-	if len(*blocks) == 0 || ((*blocks)[len(*blocks)-1].BlockHash != vLog.BlockHash || (*blocks)[len(*blocks)-1].BlockNumber != vLog.BlockNumber) {
-		fullBlock, err := etherMan.EtherClient.BlockByHash(ctx, vLog.BlockHash)
-		if err != nil {
-			return fmt.Errorf("error getting hashParent. BlockNumber: %d. Error: %w", vLog.BlockNumber, err)
-		}
-		block := prepareBlock(vLog, time.Unix(int64(fullBlock.Time()), 0), fullBlock)
-		block.Deposits = append(block.Deposits, deposit)
-		*blocks = append(*blocks, block)
-	} else if (*blocks)[len(*blocks)-1].BlockHash == vLog.BlockHash && (*blocks)[len(*blocks)-1].BlockNumber == vLog.BlockNumber {
-		(*blocks)[len(*blocks)-1].Deposits = append((*blocks)[len(*blocks)-1].Deposits, deposit)
-	} else {
-		log.Error("Error processing deposit event. BlockHash:", vLog.BlockHash, ". BlockNumber: ", vLog.BlockNumber)
-		return fmt.Errorf("error processing Deposit event")
-	}
-	or := Order{
-		Name: DepositsOrder,
-		Pos:  len((*blocks)[len(*blocks)-1].Deposits) - 1,
-	}
-	(*blocksOrder)[(*blocks)[len(*blocks)-1].BlockHash] = append((*blocksOrder)[(*blocks)[len(*blocks)-1].BlockHash], or)
-	return nil
-}
-
-func (etherMan *Client) claimEvent(ctx context.Context, vLog types.Log, blocks *[]Block, blocksOrder *map[common.Hash][]Order) error {
-	log.Debug("Claim event detected")
-	c, err := etherMan.Bridge.ParseClaimEvent(vLog)
-	if err != nil {
-		return err
-	}
-	var claim Claim
-	claim.Amount = c.Amount
-	claim.DestinationAddress = c.DestinationAddress
-	claim.Index = uint(c.Index)
-	claim.OriginalNetwork = uint(c.OriginNetwork)
-	claim.Token = c.OriginTokenAddress
-	claim.BlockNumber = vLog.BlockNumber
-	claim.TxHash = vLog.TxHash
-
-	if len(*blocks) == 0 || ((*blocks)[len(*blocks)-1].BlockHash != vLog.BlockHash || (*blocks)[len(*blocks)-1].BlockNumber != vLog.BlockNumber) {
-		fullBlock, err := etherMan.EtherClient.BlockByHash(ctx, vLog.BlockHash)
-		if err != nil {
-			return fmt.Errorf("error getting hashParent. BlockNumber: %d. Error: %w", vLog.BlockNumber, err)
-		}
-		block := prepareBlock(vLog, time.Unix(int64(fullBlock.Time()), 0), fullBlock)
-		block.Claims = append(block.Claims, claim)
-		*blocks = append(*blocks, block)
-	} else if (*blocks)[len(*blocks)-1].BlockHash == vLog.BlockHash && (*blocks)[len(*blocks)-1].BlockNumber == vLog.BlockNumber {
-		(*blocks)[len(*blocks)-1].Claims = append((*blocks)[len(*blocks)-1].Claims, claim)
-	} else {
-		log.Error("Error processing claim event. BlockHash:", vLog.BlockHash, ". BlockNumber: ", vLog.BlockNumber)
-		return fmt.Errorf("error processing claim event")
-	}
-	or := Order{
-		Name: ClaimsOrder,
-		Pos:  len((*blocks)[len(*blocks)-1].Claims) - 1,
-	}
-	(*blocksOrder)[(*blocks)[len(*blocks)-1].BlockHash] = append((*blocksOrder)[(*blocks)[len(*blocks)-1].BlockHash], or)
-	return nil
-}
-
-func (etherMan *Client) tokenWrappedEvent(ctx context.Context, vLog types.Log, blocks *[]Block, blocksOrder *map[common.Hash][]Order) error {
-	log.Debug("TokenWrapped event detected")
-	tw, err := etherMan.Bridge.ParseNewWrappedToken(vLog)
-	if err != nil {
-		return err
-	}
-	var tokenWrapped TokenWrapped
-	tokenWrapped.OriginalNetwork = uint(tw.OriginNetwork)
-	tokenWrapped.OriginalTokenAddress = tw.OriginTokenAddress
-	tokenWrapped.WrappedTokenAddress = tw.WrappedTokenAddress
-	tokenWrapped.BlockNumber = vLog.BlockNumber
-
-	if len(*blocks) == 0 || ((*blocks)[len(*blocks)-1].BlockHash != vLog.BlockHash || (*blocks)[len(*blocks)-1].BlockNumber != vLog.BlockNumber) {
-		fullBlock, err := etherMan.EtherClient.BlockByHash(ctx, vLog.BlockHash)
-		if err != nil {
-			return fmt.Errorf("error getting hashParent. BlockNumber: %d. Error: %w", vLog.BlockNumber, err)
-		}
-		block := prepareBlock(vLog, time.Unix(int64(fullBlock.Time()), 0), fullBlock)
-		block.Tokens = append(block.Tokens, tokenWrapped)
-		*blocks = append(*blocks, block)
-	} else if (*blocks)[len(*blocks)-1].BlockHash == vLog.BlockHash && (*blocks)[len(*blocks)-1].BlockNumber == vLog.BlockNumber {
-		(*blocks)[len(*blocks)-1].Tokens = append((*blocks)[len(*blocks)-1].Tokens, tokenWrapped)
-	} else {
-		log.Error("Error processing TokenWrapped event. BlockHash:", vLog.BlockHash, ". BlockNumber: ", vLog.BlockNumber)
-		return fmt.Errorf("error processing TokenWrapped event")
-	}
-	or := Order{
-		Name: TokensOrder,
-		Pos:  len((*blocks)[len(*blocks)-1].Tokens) - 1,
-	}
-	(*blocksOrder)[(*blocks)[len(*blocks)-1].BlockHash] = append((*blocksOrder)[(*blocks)[len(*blocks)-1].BlockHash], or)
-	return nil
-}
-
-func (etherMan *Client) sequencedBatchesEvent(ctx context.Context, vLog types.Log, blocks *[]Block, blocksOrder *map[common.Hash][]Order) error {
-	log.Debug("SequenceBatches event detected")
-	sb, err := etherMan.PoE.ParseSequenceBatches(vLog)
-	if err != nil {
-		return err
-	}
-	// Read the tx for this event.
-	tx, isPending, err := etherMan.EtherClient.TransactionByHash(ctx, vLog.TxHash)
-	if err != nil {
-		return err
-	} else if isPending {
-		return fmt.Errorf("error tx is still pending. TxHash: %s", tx.Hash().String())
-	}
-	msg, err := tx.AsMessage(types.NewLondonSigner(tx.ChainId()), big.NewInt(0))
-	if err != nil {
-		log.Error(err)
-		return err
-	}
-	sequences, err := decodeSequences(tx.Data(), sb.NumBatch, msg.From(), vLog.TxHash)
-	if err != nil {
-		return fmt.Errorf("error decoding the sequences: %v", err)
-	}
-
-	if len(*blocks) == 0 || ((*blocks)[len(*blocks)-1].BlockHash != vLog.BlockHash || (*blocks)[len(*blocks)-1].BlockNumber != vLog.BlockNumber) {
-		fullBlock, err := etherMan.EtherClient.BlockByHash(ctx, vLog.BlockHash)
-		if err != nil {
-			return fmt.Errorf("error getting hashParent. BlockNumber: %d. Error: %w", vLog.BlockNumber, err)
-		}
-		block := prepareBlock(vLog, time.Unix(int64(fullBlock.Time()), 0), fullBlock)
-		block.SequencedBatches = append(block.SequencedBatches, sequences)
-		*blocks = append(*blocks, block)
-	} else if (*blocks)[len(*blocks)-1].BlockHash == vLog.BlockHash && (*blocks)[len(*blocks)-1].BlockNumber == vLog.BlockNumber {
-		(*blocks)[len(*blocks)-1].SequencedBatches = append((*blocks)[len(*blocks)-1].SequencedBatches, sequences)
-	} else {
-		log.Error("Error processing SequencedBatches event. BlockHash:", vLog.BlockHash, ". BlockNumber: ", vLog.BlockNumber)
-		return fmt.Errorf("error processing SequencedBatches event")
-	}
-	or := Order{
-		Name: SequenceBatchesOrder,
-		Pos:  len((*blocks)[len(*blocks)-1].SequencedBatches) - 1,
-	}
-	(*blocksOrder)[(*blocks)[len(*blocks)-1].BlockHash] = append((*blocksOrder)[(*blocks)[len(*blocks)-1].BlockHash], or)
-	return nil
-}
-
-func decodeSequences(txData []byte, lastBatchNumber uint64, sequencer common.Address, txHash common.Hash) ([]SequencedBatch, error) {
-	// Extract coded txs.
-	// Load contract ABI
-	abi, err := abi.JSON(strings.NewReader(proofofefficiency.ProofofefficiencyABI))
-	if err != nil {
-		return nil, err
-	}
-
-	// Recover Method from signature and ABI
-	method, err := abi.MethodById(txData[:4])
-	if err != nil {
-		return nil, err
-	}
-
-	// Unpack method inputs
-	data, err := method.Inputs.Unpack(txData[4:])
-	if err != nil {
-		return nil, err
-	}
-	var sequences []proofofefficiency.ProofOfEfficiencyBatchData
-	bytedata, err := json.Marshal(data[0])
-	if err != nil {
-		return nil, err
-	}
-	err = json.Unmarshal(bytedata, &sequences)
-	if err != nil {
-		return nil, err
-	}
-
-	sequencedBatches := make([]SequencedBatch, len(sequences))
-	for i := len(sequences) - 1; i >= 0; i-- {
-		lastBatchNumber -= uint64(len(sequences[i].ForceBatchesTimestamp))
-		sequencedBatches[i] = SequencedBatch{
-			BatchNumber:                lastBatchNumber,
-			Sequencer:                  sequencer,
-			TxHash:                     txHash,
-			ProofOfEfficiencyBatchData: sequences[i],
-		}
-		lastBatchNumber--
-	}
-
-	return sequencedBatches, nil
-}
-
-func (etherMan *Client) verifyBatchEvent(ctx context.Context, vLog types.Log, blocks *[]Block, blocksOrder *map[common.Hash][]Order) error {
-	log.Debug("VerifyBatch event detected")
-	vb, err := etherMan.PoE.ParseVerifyBatch(vLog)
-	if err != nil {
-		return err
-	}
-	var verifyBatch VerifiedBatch
-	verifyBatch.BatchNumber = vb.NumBatch
-	verifyBatch.TxHash = vLog.TxHash
-	verifyBatch.Aggregator = vb.Aggregator
-	verifyBatch.BlockNumber = vLog.BlockNumber
-
-	if len(*blocks) == 0 || ((*blocks)[len(*blocks)-1].BlockHash != vLog.BlockHash || (*blocks)[len(*blocks)-1].BlockNumber != vLog.BlockNumber) {
-		fullBlock, err := etherMan.EtherClient.BlockByHash(ctx, vLog.BlockHash)
-		if err != nil {
-			return fmt.Errorf("error getting hashParent. BlockNumber: %d. Error: %w", vLog.BlockNumber, err)
-		}
-		block := prepareBlock(vLog, time.Unix(int64(fullBlock.Time()), 0), fullBlock)
-		*blocks = append(*blocks, block)
-		block.VerifiedBatches = append(block.VerifiedBatches, verifyBatch)
-		*blocks = append(*blocks, block)
-	} else if (*blocks)[len(*blocks)-1].BlockHash == vLog.BlockHash && (*blocks)[len(*blocks)-1].BlockNumber == vLog.BlockNumber {
-		(*blocks)[len(*blocks)-1].VerifiedBatches = append((*blocks)[len(*blocks)-1].VerifiedBatches, verifyBatch)
-	} else {
-		log.Error("Error processing VerifyBatch event. BlockHash:", vLog.BlockHash, ". BlockNumber: ", vLog.BlockNumber)
-		return fmt.Errorf("error processing VerifyBatch event")
-	}
-	or := Order{
-		Name: VerifyBatchOrder,
-		Pos:  len((*blocks)[len(*blocks)-1].VerifiedBatches) - 1,
-	}
-	(*blocksOrder)[(*blocks)[len(*blocks)-1].BlockHash] = append((*blocksOrder)[(*blocks)[len(*blocks)-1].BlockHash], or)
-	return nil
-}
-
-func (etherMan *Client) forceSequencedBatchesEvent(ctx context.Context, vLog types.Log, blocks *[]Block, blocksOrder *map[common.Hash][]Order) error {
-	log.Debug("SequenceForceBatches event detect")
-	fsb, err := etherMan.PoE.ParseSequenceForceBatches(vLog)
-	if err != nil {
-		return err
-	}
-
-	var sequencedForceBatch SequencedForceBatch
-	sequencedForceBatch.LastBatchSequenced = fsb.NumBatch
-	if err != nil {
-		return err
-	}
-	// Read the tx for this batch.
-	tx, isPending, err := etherMan.EtherClient.TransactionByHash(ctx, vLog.TxHash)
-	if err != nil {
-		return err
-	} else if isPending {
-		return fmt.Errorf("error: tx is still pending. TxHash: %s", tx.Hash().String())
-	}
-	msg, err := tx.AsMessage(types.NewLondonSigner(tx.ChainId()), big.NewInt(0))
-	if err != nil {
-		log.Error(err)
-		return err
-	}
-	sequencedForceBatch.Sequencer = msg.From()
-	sequencedForceBatch.ForceBatchNumber, err = decodeForceBatchNumber(tx.Data())
-	sequencedForceBatch.TxHash = vLog.TxHash
-	if err != nil {
-		return err
-	}
-
-	if len(*blocks) == 0 || ((*blocks)[len(*blocks)-1].BlockHash != vLog.BlockHash || (*blocks)[len(*blocks)-1].BlockNumber != vLog.BlockNumber) {
-		fullBlock, err := etherMan.EtherClient.BlockByHash(ctx, vLog.BlockHash)
-		if err != nil {
-			return fmt.Errorf("error getting hashParent. BlockNumber: %d. Error: %w", vLog.BlockNumber, err)
-		}
-		block := prepareBlock(vLog, time.Unix(int64(fullBlock.Time()), 0), fullBlock)
-		block.SequencedForceBatches = append(block.SequencedForceBatches, sequencedForceBatch)
-		*blocks = append(*blocks, block)
-	} else if (*blocks)[len(*blocks)-1].BlockHash == vLog.BlockHash && (*blocks)[len(*blocks)-1].BlockNumber == vLog.BlockNumber {
-		(*blocks)[len(*blocks)-1].SequencedForceBatches = append((*blocks)[len(*blocks)-1].SequencedForceBatches, sequencedForceBatch)
-	} else {
-		log.Error("Error processing ForceSequencedBatches event. BlockHash:", vLog.BlockHash, ". BlockNumber: ", vLog.BlockNumber)
-		return fmt.Errorf("error processing ForceSequencedBatches event")
-	}
-	or := Order{
-		Name: SequenceForceBatchesOrder,
-		Pos:  len((*blocks)[len(*blocks)-1].SequencedForceBatches) - 1,
-	}
-	(*blocksOrder)[(*blocks)[len(*blocks)-1].BlockHash] = append((*blocksOrder)[(*blocks)[len(*blocks)-1].BlockHash], or)
-
-	return nil
-}
-
-func (etherMan *Client) forcedBatchEvent(ctx context.Context, vLog types.Log, blocks *[]Block, blocksOrder *map[common.Hash][]Order) error {
-	log.Debug("ForceBatch event detected")
-	fb, err := etherMan.PoE.ParseForceBatch(vLog)
-	if err != nil {
-		return err
-	}
-	var forcedBatch ForcedBatch
-	forcedBatch.ForcedBatchNumber = fb.ForceBatchNum
-	forcedBatch.GlobalExitRoot = fb.LastGlobalExitRoot
-	forcedBatch.BlockNumber = vLog.BlockNumber
-	// Read the tx for this batch.
-	tx, isPending, err := etherMan.EtherClient.TransactionByHash(ctx, vLog.TxHash)
-	if err != nil {
-		return err
-	} else if isPending {
-		return fmt.Errorf("error: tx is still pending. TxHash: %s", tx.Hash().String())
-	}
-	msg, err := tx.AsMessage(types.NewLondonSigner(tx.ChainId()), big.NewInt(0))
-	if err != nil {
-		log.Error(err)
-		return err
-	}
-	if fb.Sequencer == msg.From() {
-		forcedBatch.RawTxsData = tx.Data()
-	} else {
-		forcedBatch.RawTxsData = fb.Transactions
-	}
-	forcedBatch.Sequencer = fb.Sequencer
-	fullBlock, err := etherMan.EtherClient.BlockByHash(ctx, vLog.BlockHash)
-	if err != nil {
-		return fmt.Errorf("error getting hashParent. BlockNumber: %d. Error: %w", vLog.BlockNumber, err)
-	}
-	t := time.Unix(int64(fullBlock.Time()), 0)
-	forcedBatch.ForcedAt = t
-
-	if len(*blocks) == 0 || ((*blocks)[len(*blocks)-1].BlockHash != vLog.BlockHash || (*blocks)[len(*blocks)-1].BlockNumber != vLog.BlockNumber) {
-		block := prepareBlock(vLog, t, fullBlock)
-		block.ForcedBatches = append(block.ForcedBatches, forcedBatch)
-		*blocks = append(*blocks, block)
-	} else if (*blocks)[len(*blocks)-1].BlockHash == vLog.BlockHash && (*blocks)[len(*blocks)-1].BlockNumber == vLog.BlockNumber {
-		(*blocks)[len(*blocks)-1].ForcedBatches = append((*blocks)[len(*blocks)-1].ForcedBatches, forcedBatch)
-	} else {
-		log.Error("Error processing ForceBatch event. BlockHash:", vLog.BlockHash, ". BlockNumber: ", vLog.BlockNumber)
-		return fmt.Errorf("error processing ForceBatch event")
-	}
-	or := Order{
-		Name: ForcedBatchesOrder,
-		Pos:  len((*blocks)[len(*blocks)-1].ForcedBatches) - 1,
-	}
-	(*blocksOrder)[(*blocks)[len(*blocks)-1].BlockHash] = append((*blocksOrder)[(*blocks)[len(*blocks)-1].BlockHash], or)
-	return nil
-}
-
-func decodeForceBatchNumber(txData []byte) (uint64, error) {
-	// Extract coded txs.
-	// Load contract ABI
-	abi, err := abi.JSON(strings.NewReader(proofofefficiency.ProofofefficiencyABI))
-	if err != nil {
-		return 0, err
-	}
-
-	// Recover Method from signature and ABI
-	method, err := abi.MethodById(txData[:4])
-	if err != nil {
-		return 0, err
-	}
-
-	// Unpack method inputs
-	data, err := method.Inputs.Unpack(txData[4:])
-	if err != nil {
-		return 0, err
-	}
-
-	return data[0].(uint64), nil
-}
-
-func prepareBlock(vLog types.Log, t time.Time, fullBlock *types.Block) Block {
-	var block Block
-	block.BlockNumber = vLog.BlockNumber
-	block.BlockHash = vLog.BlockHash
-	block.ParentHash = fullBlock.ParentHash()
-	block.ReceivedAt = t
-	return block
-}
-
-func hash(data ...[32]byte) [32]byte {
-	var res [32]byte
-	hash := sha3.NewLegacyKeccak256()
-	for _, d := range data {
-		hash.Write(d[:]) //nolint:errcheck,gosec
-	}
-	copy(res[:], hash.Sum(nil))
-	return res
-}
-
-// HeaderByNumber returns a block header from the current canonical chain. If number is
-// nil, the latest known header is returned.
-func (etherMan *Client) HeaderByNumber(ctx context.Context, number *big.Int) (*types.Header, error) {
-	return etherMan.EtherClient.HeaderByNumber(ctx, number)
-}
-
-// EthBlockByNumber function retrieves the ethereum block information by ethereum block number.
-func (etherMan *Client) EthBlockByNumber(ctx context.Context, blockNumber uint64) (*types.Block, error) {
-	block, err := etherMan.EtherClient.BlockByNumber(ctx, new(big.Int).SetUint64(blockNumber))
-	if err != nil {
-		if errors.Is(err, ethereum.NotFound) || err.Error() == "block does not exist in blockchain" {
-			return nil, ErrNotFound
-		}
-		return nil, err
-	}
-	return block, nil
-}
-
-// GetLatestBatchNumber function allows to retrieve the latest proposed batch in the smc
-func (etherMan *Client) GetLatestBatchNumber() (uint64, error) {
-	latestBatch, err := etherMan.PoE.LastBatchSequenced(&bind.CallOpts{Pending: false})
-	return uint64(latestBatch), err
-}
