--- synchronizerv2/config.go
@@ -1,16 +1,0 @@
-package synchronizerv2
-
-import (
-	"github.com/0xPolygonHermez/zkevm-node/config/types"
-)
-
-// Config represents the configuration of the synchronizer
-type Config struct {
-	// SyncInterval is the delay interval between reading new rollup information
-	SyncInterval types.Duration `mapstructure:"SyncInterval"`
-
-	// SyncChunkSize is the number of blocks to sync on each chunk
-	SyncChunkSize uint64 `mapstructure:"SyncChunkSize"`
-
-	GrpcURL string `mapstructure:"GrpcURL"`
-}
